GH-3178: Refactor Runner code for test manifests
